Add badges to Products and Orders tables, and implement action for toggling order invoice

+++ src/lib/utils.ts
@@ -14,6 +14,6 @@ export function formatCurrency(amount: number) {
 
 export function formatDate(date: Date) {
   return new Intl.DateTimeFormat("en-CA", {
-    dateStyle: "short",
+    dateStyle: "long",
   }).format(date);
 }
